requirements.txt
google-genai
graphviz